components to files
