+++ src/blocks/Bar.js
@@ -1,5 +1,3 @@
-import React from 'react'
-
 function Bar() {
   return (
     <div className="bar">
