+++ src/blocks/Main.js
@@ -1,0 +1,16 @@
+import React, { useState } from 'react'
+import SideBar from './SideBar'
+import Nav from './Nav'
+import CenterBlock from './CenterBlock'
+
+function Main() {
+  return (
+    <main className="main">
+      <Nav />
+      <CenterBlock />
+      <SideBar />
+    </main>
+  )
+}
+
+export default Main
